Scanner fixes and improvements, updated copyrights
In Scanner changed the principle of work with the input stream.
By default, Scanner was keeping characters. It was necessary to call
a separate command to drop the current character. But Scanner mostly
drop characters. It's more proper for Scanner to require to keep the
current character rather than especially request to drop it.

A new flag was created to implement this behavior: stopAtCurrent.
Scanner raises the flag to keep the current character.